requirements.txt
ultralytics 
torch 
opencv-python-headless
pyserial 
pyyaml 
requests
torchvision
tkinter